requirements.txt
scapy
netifaces
tftpy
termcolor
httpserver
pyyaml
Template
Jinja2